requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
python-jose
passlib[bcrypt]
bcrypt==3.2.2
python-dotenv
httpx
